step 1 of cache refactor #795; removed sys module hackery
