+++ librosa/__init__.py
@@ -8,7 +8,7 @@ from .version import version as __version__
 from .version import show_versions
 
 # And all the librosa sub-modules
-from . import cache
+from ._cache import cache
 from . import core
 from . import beat
 from . import decompose
