+++ librosa/_cache.py
@@ -3,7 +3,6 @@
 """Function caching"""
 
 import os
-import sys
 from joblib import Memory
 
 
@@ -55,11 +54,8 @@ class CacheManager(Memory):
 
 
 # Instantiate the cache from the environment
-CACHE = CacheManager(os.environ.get('LIBROSA_CACHE_DIR', None),
+cache = CacheManager(os.environ.get('LIBROSA_CACHE_DIR', None),
                      mmap_mode=os.environ.get('LIBROSA_CACHE_MMAP', None),
                      compress=os.environ.get('LIBROSA_CACHE_COMPRESS', False),
                      verbose=int(os.environ.get('LIBROSA_CACHE_VERBOSE', 0)),
                      level=int(os.environ.get('LIBROSA_CACHE_LEVEL', 10)))
-
-# Override the module's __call__ attribute
-sys.modules[__name__] = CACHE
