+++ librosa/beat.py
@@ -13,7 +13,7 @@ Beat and tempo
 import numpy as np
 import scipy
 
-from . import cache
+from ._cache import cache
 from . import core
 from . import onset
 from . import util
