+++ librosa/core/audio.py
@@ -12,7 +12,7 @@ import resampy
 
 from .fft import get_fftlib
 from .time_frequency import frames_to_samples, time_to_samples
-from .. import cache
+from .._cache import cache
 from .. import util
 from ..util.exceptions import ParameterError
