+++ librosa/filters.py
@@ -50,7 +50,7 @@ import six
 
 from numba import jit
 
-from . import cache
+from ._cache import cache
 from . import util
 from .util.exceptions import ParameterError
 from .util.decorators import deprecated
