+++ tests/test_cache.py
@@ -11,6 +11,7 @@ import shutil
 import numpy as np
 
 import pytest
+import librosa._cache
 
 
 # Disable any initial cache settings
@@ -24,9 +25,8 @@ for key in ['DIR', 'MMAP', 'COMPRESS', 'VERBOSE', 'LEVEL']:
 @pytest.fixture
 def local_cache():
     cache_dir = tempfile.mkdtemp()
-    os.environ['LIBROSA_CACHE_DIR'] = cache_dir
-    yield
-    os.environ.pop('LIBROSA_CACHE_DIR')
+    cache = librosa._cache.CacheManager(cache_dir, verbose=0, level=10)
+    yield cache
     shutil.rmtree(cache_dir)
 
 
@@ -37,26 +37,21 @@ def func(x):
 
 def test_cache_disabled():
 
-    os.environ.pop('LIBROSA_CACHE_DIR', None)
-    sys.modules.pop('librosa.cache', None)
-    import librosa.cache
-
-    func_cache = librosa.cache(level=-10)(func)
-
     # When there's no cache directory in the environment,
     # librosa.cache is a no-op.
+    cache = librosa._cache.CacheManager(None, verbose=0, level=10)
+    func_cache = cache(level=-10)(func)
+
     assert func == func_cache
 
 
 def test_cache_enabled(local_cache):
 
-    sys.modules.pop('librosa.cache', None)
-    import librosa.cache
-    librosa.cache.clear()
+    local_cache.clear()
 
-    func_cache = librosa.cache(level=-10)(func)
+    func_cache = local_cache(level=-10)(func)
 
-    # The cache should be active now
+    # The cache should be active now, so func_cache should be a different object from func
     assert func_cache != func
 
     # issue three calls to func
